requirements.txt
requests
Pillow
fastapi
uvicorn
selenium